+++ src/geografia/admin.py
@@ -1,9 +1,9 @@
 from django.contrib import admin
 
 from django.contrib import admin
-from .models import Provincia, ConsellComarcal, Municipi
+from .models import Provincia, Comarca, Municipi
 
 # Register your models here.
-admin.site.register(Provincia, ConsellComarcal, Municipi)
+admin.site.register(Provincia, Comarca, Municipi)
